+++ tests/test_address_helper.py
@@ -1,119 +1,146 @@
 import pandas as pd
 
-from helpers.address_helper import reformat_malformed_addresses
+from collections import OrderedDict
+from helpers.address_helper import reformat_malformed_address
 from pandas.testing import assert_series_equal
+from usaddress import RepeatedLabelError
 
 
 class TestAddressHelper:
 
-    def test_empty_address(self):
+    def test_good_address(self, mocker):
         input_row = pd.Series({
-            'address': None,
-            'city': '',
-            'region': '   ',
-            'postal_code': None})
-        output_row = pd.Series({
-            'address': '',
-            'city': '',
-            'region': '',
-            'postal_code': ''})
-        assert_series_equal(
-            reformat_malformed_addresses(input_row), output_row)
-
-    def test_good_address(self):
-        input_row = pd.Series({
-            'address': '123 REAL AVE',
+            'address': '123 REAL AVE APT 1',
             'city': 'NEW YORK',
             'region': 'NY',
-            'postal_code': '11111-2222'})
-        assert_series_equal(
-            reformat_malformed_addresses(input_row), input_row)
+            'postal_code': '11111-2222',
+            'full_address': '123 REAL AVE APT 1 NEW YORK NY 11111-2222'})
+        mocker.patch('usaddress.tag', return_value=(OrderedDict([
+            ('AddressNumber', '123'),
+            ('street', 'REAL AVE'),
+            ('line2', 'APT 1'),
+            ('PlaceName', 'NEW YORK'),
+            ('StateName', 'NY'),
+            ('ZipCode', '11111-2222')]),
+            'StreetAddress'))
+        output_row = input_row.copy()
+        output_row['house_number'] = '123'
+        output_row['street_name'] = 'REAL AVE'
 
-    def test_one_line_address(self):
+        assert_series_equal(
+            reformat_malformed_address(input_row), output_row)
+
+    def test_misordered_address(self, mocker):
         input_row = pd.Series({
-            'address': '123 REAL AVE NEW YORK NY 11111-2222',
-            'city': None,
-            'region': None,
-            'postal_code': None})
+            'address': '123',
+            'city': 'REAL AVE APT 1',
+            'region': 'NEW YORK NY',
+            'postal_code': '11111-2222',
+            'full_address': '123 REAL AVE APT 1 NEW YORK NY 11111-2222'})
+        mocker.patch('usaddress.tag', return_value=(OrderedDict([
+            ('AddressNumber', '123'),
+            ('street', 'REAL AVE'),
+            ('line2', 'APT 1'),
+            ('PlaceName', 'NEW YORK'),
+            ('StateName', 'NY'),
+            ('ZipCode', '11111-2222')]),
+            'StreetAddress'))
         output_row = pd.Series({
-            'address': '123 REAL AVE',
+            'address': '123 REAL AVE APT 1',
             'city': 'NEW YORK',
             'region': 'NY',
-            'postal_code': '11111-2222'})
+            'postal_code': '11111-2222',
+            'full_address': '123 REAL AVE APT 1 NEW YORK NY 11111-2222',
+            'house_number': '123',
+            'street_name': 'REAL AVE'})
+
         assert_series_equal(
-            reformat_malformed_addresses(input_row), output_row)
+            reformat_malformed_address(input_row), output_row)
 
-    def test_found_postal_code(self):
-        input_df = pd.DataFrame([
-            ['123 REAL AVE 11111', 'NEW YORK', 'NY', None],
-            ['123 REAL AVE', 'NEW YORK 11111', 'NY', None],
-            ['123 REAL AVE', 'NEW YORK', 'NY 11111', None],
-            ['123 REAL AVE 11111', 'NEW YORK 11111', 'NY 11111', 'bad']],
-            columns=['address', 'city', 'region', 'postal_code'])
-        output_row = pd.Series({
-            'address': '123 REAL AVE',
-            'city': 'NEW YORK',
-            'region': 'NY',
-            'postal_code': '11111'})
-
-        for _, input_row in input_df.iterrows():
-            assert_series_equal(
-                reformat_malformed_addresses(input_row), output_row,
-                check_names=False)
-
-    def test_found_borough(self):
-        input_df = pd.DataFrame([
-            ['123 REAL AVE BRONX', None, 'NY', '11111'],
-            ['123 REAL AVE', None, 'BROOKLYN', '11111'],
-            ['123 REAL AVE', None, 'NY', 'STATEN ISLAND'],
-            ['123 REAL AVE QUEENS', None, 'NY QUEENS', 'QUEENS'],
-            ['123 REAL AVE', 'MANHATTAN', 'bad', '11111'],
-            ['123 REAL AVE NEW YORK NY', None, None, '11111'],
-            ['123 REAL AVE NEW YORK', None, None, '11111']],
-            columns=['address', 'city', 'region', 'postal_code'])
-        output_df = pd.DataFrame([
-            ['123 REAL AVE', 'BRONX', 'NY', '11111'],
-            ['123 REAL AVE', 'BROOKLYN', 'NY', '11111'],
-            ['123 REAL AVE', 'STATEN ISLAND', 'NY', ''],
-            ['123 REAL AVE', 'QUEENS', 'NY', ''],
-            ['123 REAL AVE', 'NEW YORK', 'NY', '11111'],
-            ['123 REAL AVE', 'NEW YORK', 'NY', '11111'],
-            ['123 REAL AVE', '', 'NY', '11111']],
-            columns=['address', 'city', 'region', 'postal_code'])
-
-        for index, input_row in input_df.iterrows():
-            assert_series_equal(
-                reformat_malformed_addresses(input_row), output_df.loc[index],
-                check_names=False)
-
-    def test_found_ny_state(self):
-        input_df = pd.DataFrame([
-            ['123 REAL AVE NEW YORK', 'REAL CITY', None, '11111'],
-            ['123 REAL AVE', 'REAL CITY NY', None, '11111'],
-            ['123 REAL AVE', 'REAL CITY', None, 'NYC 11111'],
-            ['123 REAL AVE NY', 'REAL CITY NY', None, '11111 NY']],
-            columns=['address', 'city', 'region', 'postal_code'])
-        output_row = pd.Series({
-            'address': '123 REAL AVE',
-            'city': 'REAL CITY',
-            'region': 'NY',
-            'postal_code': '11111'})
-
-        for _, input_row in input_df.iterrows():
-            assert_series_equal(
-                reformat_malformed_addresses(input_row), output_row,
-                check_names=False)
-
-    def test_character_replacement(self):
+    def test_character_replacement(self, mocker):
         input_row = pd.Series({
-            'address': '1\\2"3\' $R%E{A[L∆ AVE',
+            'address': '123 $R%E{A[L∆ AVE',
             'city': 'N1E2W3 Y.O,R#K',
             'region': '1N&Y.',
-            'postal_code': 'abc11111-2.2,2+2d'})
+            'postal_code': 'abc11111-2.2,2+2d',
+            'full_address': ('123 $R%E{A[L∆ AVE N1E2W3 Y.O,R#K 1N&Y. '
+                             'abc11111-2.2,2+2d')})
+        mocker.patch('usaddress.tag', return_value=(OrderedDict([
+            ('AddressNumber', '123'),
+            ('street', '$R%E{A[L∆ AVE'),
+            ('PlaceName', 'N1E2W3 Y.O,R#K'),
+            ('StateName', '1N&Y.'),
+            ('ZipCode', 'abc11111-2.2,2+2d')]),
+            'StreetAddress'))
         output_row = pd.Series({
             'address': '123 REAL AVE',
             'city': 'NEW YORK',
             'region': 'NY',
-            'postal_code': '11111-2222'})
+            'postal_code': '11111-2222',
+            'full_address': ('123 $R%E{A[L∆ AVE N1E2W3 Y.O,R#K 1N&Y. '
+                             'abc11111-2.2,2+2d'),
+            'house_number': '123',
+            'street_name': 'REAL AVE'})
+
         assert_series_equal(
-            reformat_malformed_addresses(input_row), output_row)
+            reformat_malformed_address(input_row), output_row)
+
+    def test_repeated_labels_error(self, mocker):
+        input_row = pd.Series({
+            'address': '123 REAL AVE APT 1',
+            'city': 'NEW YORK',
+            'region': 'NY',
+            'postal_code': '11111-2222',
+            'full_address': '123 REAL AVE APT 1 NEW YORK NY 11111-2222'})
+        mocker.patch('usaddress.tag', side_effect=RepeatedLabelError(
+            input_row['full_address'],
+            [('123', 'AddressNumber'),
+             ('REAL', 'StreetName'),
+             ('AVE', 'StreetName'),
+             ('APT', 'OccupancyType'),
+             ('1', 'OccupancyIdentifier'),
+             ('NEW', 'PlaceName'),
+             ('YORK NY', 'PlaceName'),
+             ('11111', 'ZipCode')],
+            'StreetAddress'))
+        output_row = pd.Series({
+            'address': '123 REAL AVE APT 1',
+            'city': 'NEW YORK NY',
+            'region': 'NY',
+            'postal_code': '11111',
+            'full_address': '123 REAL AVE APT 1 NEW YORK NY 11111-2222',
+            'house_number': '123',
+            'street_name': 'REAL AVE'})
+
+        assert_series_equal(
+            reformat_malformed_address(input_row), output_row)
+
+    def test_repeated_address(self, mocker):
+        input_row = pd.Series({
+            'address': '123 REAL AVE',
+            'city': '123 REAL AVE',
+            'region': 'NEW YORK NY',
+            'postal_code': '11111-2222',
+            'full_address': ('123 REAL AVE 123 REAL AVE NEW YORK NY '
+                             '11111-2222')})
+        mocker.patch('usaddress.tag', side_effect=RepeatedLabelError(
+            input_row['full_address'],
+            [('123', 'AddressNumber'),
+             ('123', 'AddressNumber'),
+             ('REAL AVE', 'StreetName'),
+             ('REAL AVE', 'StreetName'),
+             ('NEW YORK', 'PlaceName'),
+             ('NY', 'StateName'),
+             ('11111-2222', 'ZipCode')],
+            'StreetAddress'))
+        output_row = pd.Series({
+            'address': '123 REAL AVE',
+            'city': 'NEW YORK',
+            'region': 'NY',
+            'postal_code': '11111-2222',
+            'full_address': '123 REAL AVE 123 REAL AVE NEW YORK NY 11111-2222',
+            'house_number': '123',
+            'street_name': 'REAL AVE'})
+
+        assert_series_equal(
+            reformat_malformed_address(input_row), output_row)